BurstList.py
A program to determine the length of bursts of zero’s in a list of N numbers stored in a List.  Records the length of the bursts of zero’s in another List. The program should then prints the lengths of the bursts.  Using this second List with the lengths of the bursts, it also computes and prints the following: the average length of the bursts, the minimum burst length, the maximum burst length, and the total number of zeros (the sum of the burst lengths). 

Example data and results:

N (entered by the user and stored in N): 15

List of numbers (entered by the user and stored in list):

1, 0, 0, 0, 0, 3, 7, 0, 0, 0, 0, 0, 0, 50, 0

List of bursts (computed by your program and stored in BurstList): 4, 6, 1

Average burst length (computed by your program): 3.67

Minimum burst length (computed by your program): 1

Maximum burst length (computed by your program): 6

Total number of zeros (computed by your program): 11

Results:

Burst   Length

1             4

2             6

3             1

Average burst length: 3.67

Minimum burst length: 1

Maximum burst length: 6

Total number of zeros: 11